LTD-239 Reports: OGL Summary
Introduce a reports app with the ability to generate csv versions of sample
report queries. In the long term these reports will be run in data workspace but
for now it is useful to keep the code for the sample SQL queries somewhere so
the can be used in future for reference.