Update README and docker-compose
This commit adds elastic search and redis to the docker compose file as we now
require these services to run the lite application.

Elastic search is used by lite-api and redis is used by lite-frontend in
caseworker mode.

The reason redis has been added here is that both compose files
share the same lite network so it can go in either the lite-api compose file or
the lite-frontend file. Since redis might be used as a cache backend for all
layers of the app including lite-api eventually it might as well be here.