Fix flagging rules when reviewing goods
When goods are reviewed and CLC entry updated, the updated entry was not saved.
Because of this the rule gets applied to an incorrect value hence causing this issue.

Also updated rules to correctly determine groups and child entries especially when
the rating has multiple levels. In this case each sub-parent is to be considered as
group as well as individual entry.

Add helpers functions that provide list of parent nodes for a given rating or given
a group rating they provide the list of child entries.