Fix DenialMatchOnApplication  migrations
This model has a foreign to the Denial object; when this model is created the
type of primary key of Denial object is int but it was later on updated to uuid.
To fix this initially the foreign key is removed and it is added back again.
The dependencies in the migrations are modified.