Fix per queue SLA update logic
Currently we use CaseAssignment instances to update SLA on per queue basis
but these are only used when the case is assigned to a user instead it should
be the M2M field queues in Case model. Recently we added a new model based
on the through table for this M2M field so use that model instance instead.