Ignore case when looking for ExporterUser using email
If the user uses email in uppercase when registering an organisation then it
is saved as lower case when saving the BaseUser instance because of this we
fail to authenticate the user when logged in with the email in upper case.
Fix this by ignoring the case when looking for the exporter.

Add a test that ensures user can be authenticated by using email in uppercase.