Expose organisation documents in good details
If some of the organisation specific documents (RFD certificate, section5 certificate)
are uploaded by the user then the system remembers them and skips the upload when
a product is being added to the application but because of this some of the fields
in the product detail page are shown as empty/None. In this case we go through the
organisation documents and display the correct data for this purpose expose the list
of documents held against the organisation.