Add controlled status to ControlListEntry model
When some entries are decontrolled then they shouldn't be made available
for the user to select. Currently in FE we cannot differentiate whether they
are controlled/decontrolled so add the status to filter them accordingly.

Also update the seeding logic to not consider the rating text when creating
new entries as the text might also change which can potentially try to create
duplicates.