Limit ARS updates to specific product on application
Annual report summary (ARS) is specific to a product on application. If the same product
is used on other application then updating ARS in one application was updating the
value in the other application also. This has been fixed but the fix was failing
if the same product is added twice in the same application.

This is because API receives list of underlying product ids (Good) and we use them to
retrieve all matching products on application (GoodOnApplication). This is the current
behaviour because if the underlying product is same for all products on applications
then reviewing one good CLC entries automatically marks all of them as verified.
This is correct but it causes issues with Report summary.

This is fixed by taking the current product on application as input and limiting the
updates for this field only to this object.

An Audit log is also created whenever ARS value is updated.