Fix products order in generated licence template
The generated licence document (pdf) contains the list of products in the licence
that require a licence along with their name, quantity etc. Currently these
are being ordered using their description field. Because of this the order in
the licence doesn't match with the order in which they are added to the
application. This has serious implications wrt HMRC integration.

When the licence details are sent to HMRC we send the product details as per
the order they are entered in the application but when products are declared
at the customs they use the order defined in the licence. This can result in
differences in quantities and the customs system may reject the products.
Eg if the quantity shows up as higher than allowed in licence.

Fix this by removing the ordering by description and add a test to ensure
the order as per application is maintained in licence also.