Add fields to save serial number options
Currently we have these fields but the field has_identification_markings is a boolean
but in the new form there are three options. The safe option is to add new field,
update these with data from current field and finally delete the old field.

This PR just adds the new fields.