Add display_value field to DenialReason to display in UI
Some of the refusal criteria items have a descriptive text to display instead of
an acronym used in id. Changing the id is not an option as we will endup creating
new items when the seed command is executed.

This PR adds a new field to the model which is used to display in the UI instead.
For those items where there is no descriptive text, we use the value of id itself.