Update Party model to include new fields for End-user
In the journey of adding an End-user we are capturing new details.
We are asking whether there is an end user document available or not,
if it is not in english and that atleast one of the page in the
document is on company letterhead etc. These changes are specific
to End-user only and not applicable to other parties such as
Consignee, third parties etc