Return Party documents that are still being processed
The process in this case is the virus scan which is performed on every
document that is uploaded. Once the scan completes we mark them as safe.

Previously ignored returning the documents whose virus scan is in progress
but because of this we cannot show the user whether the document is uploaded
or not. By exposing them we can display the document name assuring the user
that the document is indeed uploaded and we only allow them to download
once the virus scan is completed and we know document is safe.