Update Party document validation during application submission
When exporter is ready to submit we perform various validations to check
whether all the required details are captured or not.

One such validation is to check whether Party related documents are
uploaded or not. Current validation assumes each party only uploads
one document but with English translation and letterhead changes an
End-user may upload upto 3 documents so update the validation to
take this into account.