Capture product differences note when adding End-user details
When uploading the end-user undertaking document we ask the user if there
are any differences in the products listed in the document and the products
in the application. Previously this was incorrectly associated as document
description but infact it is a separate piece of detail we are capturing
from the user.