Skip sending licence details to hmrc when there are no licenceable goods
Some of the licences contains only NLR goods, currently we are trying to
inform lite-hmrc in this case also cauing the task to fail and throwing
alerts which in this case are not errors.

If there are no licenceable goods on the licence then skip sending these
details to lite-hmrc.

v2: Add new test for this case.
Also this broke few tests because the failing hmrc related tests were not
associating the goods to the licence. Also marks two tests related to
open general applications as we don't support them yet.