Add audit event to show product review and report summary updated details
Currently we emit two separate audit events to show that a product is reviewed
along with its old and new control list entries and also old and new
report summary values. These are also shown in a way which is not readable.

Create new event to combine these two logs and present them in a readble way.
The existing event cannot be used as this requires a new piece of data (line item)
which would be missing from existing events and will give a 'KeyError' hence
a new event is created. This would mean only new cases will use this event and
existing cases display them using two separate events.