Retain previously assessed CLEs when a product is re-used in another application
When a new product is added to an application Exporter can specify CLE for
that product. This may not be the correct one depending on technicalities
and the parties involved for that application so caseworker may/can correct
this as part of their assessment.

Currently if the same product is re-used in some other application we are
replacing previous CLEs with the most recent assessment values.

The requirement is to show all of the previous assessment along with the
recent assessment also to the Exporter hence instead of overwriting we
add to existing values in the case where product is re-used.