Reset products verification status when an application is copied
When we create new applications by making a copy of an existing application
currently we are retaining the product's verification status. Because of this
when this application is submitted, the products are already shown as
verified without TAU assessing them. Fix this by resetting the verification
status when an application is copied.