Return parent node if there are no child entries
We have a helper function that returns all parent nodes of a given
CLE rating but if the given rating itself is of a parent then it
doesn't return anything. This makes flagging rules fail if that
parent is specified as matching_group, in which case rule doesn't
get applied.