Fix broken tests
We added Config data migrations in lite_routing and this made few tests fail as
the routing rules, flagging rules added in those migrations also triggered and
adding more flags to the cases in tests so they need to be updated.

Some of the factories also need updating as we are creating these in migrations,
eg Department.