Add migration to retrospectively add missing Case UPDATED_STATUS event
Audit UPDATED_STATUS was removed as part of Notes and timeline changes
and for sometime this event is not emitted when a case is finalised.
Reporting relies on this event so we added the event back but for the
missing data this migration retrospectively adds for the cases finalised
between the dates.

v2: Change the actor to be of System user to specify that this is not
done by any case officer but LITE itself.