Don't clear flags on good when TAU gives their assessment
When TAU gives their assessment then any flags set on good are cleared.
There is no reason specified but it is done when the given rating is
different from the rating specified by Exporter.

Clearing existing flags has not been an issue as the teams before TAU
have not applied any flags but this is not true with upcoming routing
changes. If these are cleared then the Case doesn't get routed correctly,
hence don't clear these flags.