Invoke routing post process function
In the post process function we handle special cases where we may have to
adjust the queues the case is going to be routed to.
Eg a situation where it is routed to both MOD-DI queues.

This has implications in terms of editing audit events. Currently we emit
these events as we process the rules so after post-processing if some queues
are disappeared then we would've already emited events for those queues.
Hence keep track of the newly applied rules and only emit events to those
queues the case is still present on.

Since we run routing rules at every status change, calling this again is
not going to be an issue as we only look at newly applied rules.