Add management command to truncate End-user name
HMRC spec only allows 80 chars for End-user name but currently we don't
have validation in UI or API to prevent this. We are only validating in
lite-hmrc when licence details are sent to be forwarded to HMRC.

Because of this we failed to send few licence details. Using this mgmt
command we can resend those details to lite-hmrc until we add these
restrictions in the UI and API.