LTD-3239: logic in view to add the selected_tab filter
The frontend cases view is going to have several 'tabs' that are meant
to only show certain cases.

The selected tab name is indicated under the `selected_tab` param. This
commit adds logic to look at the value of that param, and assign True to
the tab's name, so that it's easier for the `manager` class to create
query logic for the different filters required.