Check LU countersigning requirements and validate countersignature when finalising
A Case may need countersigning from LU depending on case attributes.
LU countersigning flags get applied to the case to indicate this. These
flags also block finalising to ensure necessary countersignatures are present
before the Case can be finalised.

In the current process countersiging is done by adding a case note and these
flags are manually removed to allow finalising the case.

In the new process countersigning is done using routing changes and we can
avoid the manual step of removing the flags by checking if necessary
countersignatures are present at the time of finalising it.

Some Cases need double countersigning so we check for both orders of
countersigning advice for this case.