Address review comments
Previously countersigning flags were not removed after ensuring
countersignatures were present. Because of this the changes were split
in multiple places as we needed to remove them from blocking flags.
During review we agreed to remove these flags which simplifies it.

Also check if any countersign advice available for refuse advice type,
this is unlikely but if the case is routed unexpectedly it might
happen so raise error in this case.

Update tests to cover all cases.