LTD-3477: add db test for `read_only` case statuses
we recently wanted to change the `under_review` status to read only, and
assumed that it was just a matter of changing the CaseStatusEnum, but it
turned out to need a migration on the CaseStatus model.

There are tests in `test_edit_application` that provide more integration
testing for both `read_only` and `editable` cases, but these make sure
that the Enum will always match what's on the database.