Add field to CountersignAdvice to mark advice as valid/invalid
Countersignatures when they are created are valid but if LU caseworker
edits their recommendation following a rejection then we need to mark
existing countersignatures as invalid. This is to differentiate from
the new countersignatures when it is sent for countersigning again.

A Case is considered countersigned when they are all approved so any
rejections can be considered as invalid when user edits their recommendation.

This will also help in the case sending the case back multiple times as
technically a Case can be sent back any number of times.

We are only going to mark as invalid instead of deleting as we want to
retain the history so updating model with this field.