Remove code that deletes existing objects when final advice is updated
We have a new put method to update final level advice and there is no
need to delete existing objects which was required for the previous
method of update (where existing objects are deleted and new ones
are created in their place).

Also update a test to ensure there is no regression by removing
this code.