Remove failing test
This test is currently failing with the LU Countersign feature flag on
as it gets stuck in infinite loop

Without countersigning changes 'under_final_review' was the last status
before finalising so the Case progresses upto that point and the loop
breaks but with Countersigning changes it progresses further to the
new statuses that we created and wraps back to 'under_final_review'
for finalising the Case. As this process happens repeatedly it is
getting stuck in infinite loop.

We identified two options,
- Update the test by adding a rule that triggers for this status
- Delete the test completely

Deleting the test completely because we have other tests that cover
first item.