Ignore deleted parties when determining Case parameter_set
Case parameter_set() is used to check for countersigning flags and if
they are present we ensure necessary countersignatures are present when
the case is finalised. However it currently doesn't ignore any deleted
parties on the application. If that party destination is the only one
that needs countersigning then can block finalising even though that party
is deleted.

Ignore deleted parties as they are no longer valid on the application.