Remove some code related to unsupported application types
Currently we only support SIEL types but some of the helper functions
account for supporting other licence types OIEL, F680, Exhibition clearance
etc so if we change those function we need to maintain these additional
types as well. This is unnecessary and can be removed.

In future if we add support for these licence types then we probably
do it in a different way so removing the code where possible and
associated tests as well.

Looking to do this in multiple stages instead of one big PR.
All changes included in one commit so it is easy to revert in case
if we need to do that.