Reset appeal_deadline when Refusal letter is generated
Exporters can appeal a refusal decision within 28 days from the date
refusal letter is generated so when the letter is actually created
we populate appeal deadline date. This is used on Exporter dashboard
to provide an option to appeal.

It is possible to regenerate refusal letters so the deadline is reset
in this case as well.

Previously we were trying to reset during the document preview but the
appropriate place is when the letter is actually created in the database.