LTD-4115: Add appeal_deadline field (#1544)
* Add field to save appeal deadline

Exporter can appeal a refusal before a certain date from the time a refusal
letter is generated.

* Update Application serializer to include appeal_deadline field

* Set appeal_deadline when a refusal letter is generated

This is reset if the letter is generated again as per requirement.

* Appeal deadline address review comments

 - Reset appeal deadline when application is finalised. This makes it
   obvious from the process point of view. Previously it was done
   as part of document generation which is not obvious as it is more
   like a side effect

 - In case of refusal we ensure this is populated before Exporter is
   notified of refusal

 - Update tests to check appeal deadline is populated correctly in
   case of refusal

* Remove unused imports in tests

* Reset appeal_deadline when Refusal letter is generated

Exporters can appeal a refusal decision within 28 days from the date
refusal letter is generated so when the letter is actually created
we populate appeal deadline date. This is used on Exporter dashboard
to provide an option to appeal.

It is possible to regenerate refusal letters so the deadline is reset
in this case as well.

Previously we were trying to reset during the document preview but the
appropriate place is when the letter is actually created in the database.