Determine advice type when regenerating document
When the final Case documents (SIEL licence or refusal or NLR letter) are generated
the advice_type is provided as a parameter. This is used to make few decisions
when generating the document.
Eg in approval case we look for draft licence and in refusal case we use this
to reset appeal deadline.

However when we regenerate document this is not available with FE and cannot
provide so look at the decisions associated with the document template and
derive this detail.