Add management command to remove Refusal advice given by LU
In the situation where an Appeal is overturned we need to prepare the
Case for LU to issue a licence. This can only be done after removing
the initial refusal advice given by LU.

Add management command that removes refusal advice. Once this command
is run then LU can finalise the Case again and issue licence.

It also moves the case to 'LU Post-Circulation Cases to finalise'
queue and removes any countersigning flags.

A dry run option is included to test before we can commit changes.