Use approved licence goods quantity and value to generate Licence document
LU have an option to edit quantity, value when finalising the Case to issue
a licence. These are updated in 'GoodOnLicence` objects for that licence.
However when licence document is generated we are pulling quantity/value
from `GoodOnApplication` which are values they applied for.

Fix this by using goods on licence.

Note that the message we sent to HMRC uses `GoodOnLicence` so it will be with
correct quantity values.