Stop marking Licence as exhausted when products are completely used
HMRC sends usage data when Exporter ships products that are licenced.
LITE keeps track of the usage and when it reaches the allowed quantity
we currently mark the licence as exhausted and inform HMRC to
cancel the licence.

Because of an issue in how we update usage we sometimes double count
and cancel the licence prematurely. Also HMRC is source of truth and
they won't allow Exporters to ship more than the quantity they are
licenced for so LITE should not really be doing this.