Remove task that cancels Licences on expiry
SIEL licences issued have an expiry date. We have task that checks periodically
for expired licences and sends a message to HMRC to cancel the licence.

HMRC is the actual source of truth and they also check whether a licence is
valid when Exporter ships goods so LITE should not really be doing this.