Allow LU user to set finalised status
To support Appeals process we need LU users to be able to set the status
to 'Finalised' at the end of an appeal. Currently this cannot be done so
update endpoint to allow only LU users with necessary permissions to
be able to set status to 'Finalised'

Using this opportunity to refactor some of the checks into a helper function.

Add test that checks only LU can set status to finalised and fails for
other internal users.