Update GoodOnApplication to save product assessor details
The Case adviser details who assessed a product are currently not saved.
These are logged in an audit event but not included as part of data model.

When users searching for historic product details they are interested in
previous assessment and advisor who made that assessment. Retrieving these
details from Audit event is not ideal as that is not the intended purpose.
It is also not reliable because Cases can come back to TAU at any stage.

Update GoodOnApplication model to save assessment date and the user.