Explicitly specify the fields to be searched
When searching across multiple fields we have an option to use copy_to
argument which copies values of fields to be searched to another field
and search is performed on that field (instead of all fields). But with
this approach we cannot define the behaviour if there are multiple
search terms eg whether we match all words or match any etc.

To help with that explicitly specify the fields to be searched.