Export worksheets of report_summary and associated fields.
This exports a spreadsheet of GoodOnApplication that don't have associated report_summary_prefix and report_summary_subject fields.
It contains suggestions for the values of fields by looking up existing values where the normalised version of report_summary matches
another GoodOnApplication containin these values.

This adds xlsxwriter to the dev requirements - this may not be needed in the final form, as we may not need the sheet once the review is done,
xlsx writer dependencies are those from within python (xlsx being xml and zip files), so if we did add it it's fairly safe.

`$ python manage.py goods_on_application_reports_sheet`

Generates the spreadsheet.

Mappings sheets:
The first sheet matches items with a subject and prefix, the second is for remaining items that don't have a prefix.

Unmapped data sheets:
The final sheets point us at data that requires a normalisation step: fixing up data here will lead us to more mappings, and should be addressed
before applying the recommendations from the first two sheets.

The first sheet is a list of GoodOnApplications and their report_summaries.

This helps us see the individual records, but there is a lot here.

The next sheet helps points us at general mappings needed: it's an alphabetical list of unique combinations of report_summary, prefix, subject.
Here we can see where spellings can be corrected or if there are any new ReportSummaryPrefix or ReportSummarySubjects need adding.