Revert "Consider nested fields also when constructing search query"
This reverts commit 0e2995ffb4d1745e13907c2776c272b67a4549ee.
Full text search with operators are not working with nested fields
as expected so we are going to flatten these fields to top level fields
and use in full text search. These are mainly CLEs and regime entries
which can be more than one for a given product so we are going to
flatten them into an array. Because of this we don't need to include
nested queries hence removing this option.