Validate search term when using query_string queries
query_string queries are strict and does not tolerate syntax errors in
the search terms. eg a trailing AND, OR operators.

In this case we get an exception when the query is executed. To avoid
this and present a meaningful error response to the user, validate the
query before executing and bailout early if the query is not valid.