Adding another data migration like 0075 revealed that it broke the tests there, fix that [see below].
The migration tests reset by running ALL migrations, unfortunately this meant that the mechanism to override the CSV
loading location, meant to be used for one test (0075) was active when migrating 0076 during the reset part of this
test, that in turn caused a failure because there was no CSV to load.

The Fix:

Instead of overriding a setting to change CSV file loadding location, use patch() during the initial migration on
Migration.get_csv_path.

This was helped by the realisation that the migrator puts the migration into sys.modules, allowing ordinary mocking.