Add product search endpoint to provide suggestions
Query string filter backend allows users to specify data category
when performing a search. Eg to search within product name then
we can type 'name:<search_term>' so that the search_term is only
looked in 'name' field.

This is not user-friendly as user needs to know the names of available
fields. Providing suggestions for the matching fields can be helpful;
as the user starts typing we can run a query and provide suggestions
as to where that search term is found. User can simply select the field
they are interested in from the suggestions provided.

Add enpoint to run query and return available suggestions for
the given search term.