Use only most recent word to look for suggestions
Once we start showing suggestions users can either choose to select
a suggestion or continue without selecting anything. In case if there
are no suggestions for that word then there is nothing to select from.

Now when they start typing second word it is expected to show suggestions
for that word only rather than the complete phrase. Hence it is important
to only consider the most recent word to retrieve suggestions.

If they indeed select a suggestion then the script resets the query string
so this case is already covered.