Increase the total number of suggestions returned
We are now matching anywhere in the string so this can return good
number of suggestions especially when only few characters are entered.

We also have many fields where we look for suggestions so to avoid
missing out on useful suggestions increase the count to 15.

This can be refined later based on user feedback.