Use complete query word for suggestions
Previously we were receiving complete search text to retrieve suggestions.
This could be multiple words so we were selecting last word out of that
to query for suggestions. This is because uses can choose to ignore suggestions
and continue to type so when they are entering second word it means either
they want to add more search terms or it could be an operator. So when
they start typing second word then we have to present suggestions only for
the second word. It is also unlikely for any suggestions to be available
for the whole phrase.

All of this is handled in FE and we only receive the word for which suggestions
are required hence use it directly.