Match phrase prefix to get suggestions
Expectation is to match on multiple words when providing suggestions.
For the multi_match query phrase_prefix type allows us to match on
multiple words and the last partial word is used as a prefix.

This can be an issue if user enters operators but the usual usage
is to first select all data categories using suggestions and then
enter operators as required in between.