Update search signals to refresh Product document also
Updates the various models are intercepted in save() signal to
refresh the indexes however it is only updating Application document.

When an application is first submitted by Exporter we run flagging rules
which result in saving of underlying 'Good' instances. At this point
the ProductDocument is not refreshed so if we search for a product in
this application we see the reference_code as None as if it is still
in draft state. Updating product document here pulls inner document
for the Application with the updated reference code.

Without this change also eventually the reference code gets refreshed
as the application progresses through various teams which result in
updating of product attributes but if it is in submitted state then
it doesn't get updated.