Add management command to create test applications
This command can be used to seed local database with certain number
of test applications.

It first creates a configurable number of product catalogue and
organisations and creates applications with random number of
products (selected from the catalogue) and submit these applications.

Note that it deletes all of existing data before it creates new data.

It also adds product assessments with random ARS subject and prefix
values. It makes certain assumptions on the probablity of prefix vs
subject values.

It is not perfect in terms of creating an perfect application but can
quickly seed specified number of applications which can be handy for
local env.