Add checks to ensure command can only be executed in local env
As this command first deletes all data before creating new test data
we need additional protection in terms of where it can be executed.

If it is run on production by mistake then that it will result in
data loss. We also have no need for this to run on a dev env.
So add condition such that it only runs on localhost.

Further ensure that this is only allowed to run when DEBUG=True
which will never be the case in production.