Add testing app to include management commands useful for testing
We need a management command to seed local database with test application
data. This command initally clear local data before creating new data.

If this is executed on any env, prod this can result in data loss so
we need to run this only when DEBUG is enabled.

The command only executes if DEBUG is enabled but isolating this in
a separate app is for additional safety. This app is only installed
when DEBUG=True which will never be the case in production.