Delay scheduling of send_licence_details_to_lite_hmrc by 10 sec
When a licence is issued many other things happen at that time which
include,
 - updating of Case status
 - sending notification email to Exporter
 - issuing of licence

When the task to send licence details to lite-hmrc runs it first
attempts to acquire lock on the table as it needs to update table.
However it is observed that the task is not able to aquire the lock.

One possibility is the updates are still pending so to avoid any
issues acquiring lock defer the task execution by 10 sec which
should be more than sufficient to flush all pending updates.