Remove check identified by codecov as it is not reacheable
Codecov identified a line which is not covered by tests. It is an
indication that this line is not reachable as we don't support any
applications other than standard. There are additional checks
that identify this case and fail earlier if we get a type other
than standard so removing the check completely.