Add unique constraints to ReportSummary model
We are not going to create new instances of ReportSummary each time a
product is assessed because there are predefined set of prefix and subject
values so can just create all possible ReportSummary instances upfront and
link them in other models.
As the combination of prefix and subject is unique we can enforce this
on db level.