Add test for invalid combination of report summary prefix and subject
ReportSummary instances are created based on all possible combinations
of prefix and subject so if we get an invalid combination we raise a
validation error. We are not likely to get invalid combination as this
are selected by user from a pre-defined list however coverage complained
that this line is not covered so add a test to check the same.