Update part number search to look for exact matches
Currently part number search in products uses a delimiter '-' when
analyzing the search string. Because of this if a part number contains
hyphens then it splits into multiple tokens and matches on individual
tokens as well. This can bring multiple results and not possible to
get an exact match.

Users are expecting for an exact match so consider the whole search
string without using any delimiters.