Back populate multiple report_summaries in GoodOnApplication instances
Using the populated prefix_id and subject_id we back populate report_summaries
field in all GoodOnApplication instances.

The migration looks at all the controlled Goods on application and populate
their report_summaries. As there can be large number of entries, instead of setting
on each good on application instance we create entries in bulk on the through table
for GoodOnApplication and ReportSummary