Update test to consider report summaries with valid prefix
Prefix is optional so it can be None, in such cases when we convert this
to string while preparing data we end up providing 'None' and the test fails
as it is not a valid UUID.

To avoid this ignore report summary objects where prefix is empty.

We prepare data such that first good has both prefix and subject, second one
is subject only and third one is NLR so ignoring objects where prefix is empty
is not an issue.