Add Trackable mixin to expose helper methods for handling model versions
With reversion we have the ability to track model changes whenever value of
any field is updated. However reversion doesn't provide any methods that
can give a history of changes for a particular field.
We can retrieve all revisions but the value of the field we are interested
in might have remained constant for multiple revisions and there is no
method available to retrieve only the version where this field value changed.

The idea is to expose such helper methods in this mixin and can be used by
all models that are registered with reversion.