Allow editing of submitted goods
Currently a good can only be edited when it is in draft state.
Once it is added to an application and that application is submitted
then the good status becomes submitted and we don't allow editing
of these goods.

But for archiving/restoring goods we need to be able edit these goods.
Only a good that is submitted/verified can be archived.

For this reason remove this restriction.
Update the existing test to check the same rather than deleting it.