Add dedicated endpoint to archive/restore a good
Exporter has a choice to archive/restore a good and this is
only allowed for goods that are submitted/verified on an application.

However we have a restriction that goods cannot be edited once they
are submitted on an application. Removing this restriction can
cause various issues so to avoid this add a dedicated endpoint
which only allows archiving/restoring a particular good.

Add tests to ensure that only archive status can be updated using
this endpoint.