Use filter backends to filter list of applications
The endpoint that returns list of applications has a requirement to
filter based on application state and user permissions. Current code
uses if-else conditions for various states which is not standard approach
and includes most of the business logic in the view.

Django filter backends can solve this in a better way and can be reused
in other views.

Add filter backends to filter based on application state and organisation
permissions.