Add exporter specific endpoint to update quantity and value
Once a product is added to an application then we are providing
option to only edit quantity and value. Currently it uses a common
endpoint using which we can also update other fields. This is
also common to caseworkers.

Add a separate endpoint which only exporter can access and allows
updating of quantity and value only.