Display only the CLEs assessed for this application on the Licence
Currently we are using CLEs from Good when displaying them on the
licence document. Usually this will be the same as recorded in
GoodOnApplication but if the underlying good is re-used in other
applications where it is assessed differently then they both
will differ. In such cases we will be displaying these values
from both the assessments.

Fix this by using the entries from GoodOnApplication.

Add test to ensure correct values are displayed.